pyproject.toml
[tool.poetry]
name = "mimic"
version = "0.1.0"
description = "Mimic lets you securely deploy web applications at scale"
authors = ["Skylar Kelty <skylar.kelty@kcl.ac.uk>"]
license = "MIT"
readme = "README.md"
packages = [{include = "mimic"}]

[tool.poetry.dependencies]
python = "^3.8"
dnspython = "^2.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = 'mimic.run_server'
client = 'mimic.run_client'
